fix: Fixed issue where delegates with the same prefix could not have output pins
notes: this was due to rather brittle logic of checking whether the
pin's name started with one of the Delegates. If the Delegates are named
such that they have the same starting name (e.g. MyDelegate and
MyDelegate2) then the ensure would trigger on the incorrect logic of
trying to find a pin on the wrong delegate